Cont: 감정기록 전역상태 등록 및 저장
