+++ src/pages/emotion/message/EmotionMessage.tsx
@@ -1,6 +1,10 @@
 // 진욱
 
+import useEmotionStore from '@store/useEmotionStore';
+
 const EmotionMessage = () => {
+  const emotionRecord = useEmotionStore((state) => state.record);
+  console.log(emotionRecord);
   return <section>나와 상대방에게 한 마디</section>;
 };
